Flux3D

Flux3D.NormalizePointCloud

Flux3D.NormalizeTriMesh

Flux3D.OffsetTriMesh

Flux3D.PointCloud

Flux3D.ReAlignPointCloud

Flux3D.ReAlignTriMesh

Flux3D.RotatePointCloud

Flux3D.RotateTriMesh

Flux3D.ScalePointCloud

Flux3D.ScaleTriMesh

Flux3D.TranslateTriMesh

Flux3D.TriMesh

Flux3D.GBMesh

Flux3D._load_meta

Flux3D._packed_to_padded

Flux3D.compute_faces_areas_list

Flux3D.compute_faces_areas_packed

Flux3D.compute_faces_areas_padded

Flux3D.compute_faces_normals_list

Flux3D.compute_faces_normals_packed

Flux3D.compute_faces_normals_padded

Flux3D.compute_verts_normals_list

Flux3D.compute_verts_normals_packed

Flux3D.compute_verts_normals_padded

Flux3D.gbmeshes

Flux3D.get_edges_packed

Flux3D.get_edges_to_key

Flux3D.get_faces_list

Flux3D.get_faces_packed

Flux3D.get_faces_padded

Flux3D.get_faces_to_edges_packed

Flux3D.get_laplacian_packed

Flux3D.get_verts_list

Flux3D.get_verts_packed

Flux3D.get_verts_padded

Flux3D.load_trimesh

Flux3D.normalize

Flux3D.normalize

Flux3D.normalize!

Flux3D.normalize!

Flux3D.npoints

Flux3D.offset

Flux3D.realign

Flux3D.realign

Flux3D.realign!

Flux3D.realign!

Flux3D.rotate

Flux3D.rotate

Flux3D.rotate!

Flux3D.rotate!

Flux3D.sample_points

Flux3D.save_trimesh

Flux3D.scale

Flux3D.scale

Flux3D.scale!

Flux3D.scale!

Flux3D.translate

Flux3D.translate!

Flux3D.NormalizePointCloud — Type

NormalizePointCloud(; inplace::Bool=true)

Normalize PointCloud with mean at origin and unit standard deviation.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud.

See also: normalize, normalize!

Flux3D.NormalizeTriMesh — Type

NormalizeTriMesh(; inplace::Bool=true)

Normalize TriMesh with mean at origin and unit standard deviation.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: normalize, normalize!

Flux3D.OffsetTriMesh — Type

OffsetTriMesh(offset_verts::AbstractArray{<:Number,2}; inplace::Bool=true)

Add offset to the TriMesh by given offset vertices offset_verts

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: offset, offset!

Flux3D.PointCloud — Type

PointCloud

Initialize PointCloud representation.

points should be Array of size (D, N, B) where N is the number of points, D is dimensionality of each points (i.e. D=2 or D=3) and B is the batch size of PointCloud. normals is optional field, if given should be Array of size (D, N, B) where N and B should match with the N and B of points and D=2 or D=3 (i.e. normals for 2D and 3D PointCloud respectively).

Fields:

points - Points that makes up whole PointCloud.

normals - Normals of each points in PointCloud.

Available Contructor:

PointCloud(points, normals=nothing)

PointCloud(;points, normals=nothing)

PointCloud(pcloud::PointCloud)

Flux3D.ReAlignPointCloud — Type

ReAlignPointCloud(target::PointCloud; inplace::Bool=true)
ReAlignPointCloud(target::AbstractArray{<:Number, 2}; inplace::Bool=true)

Re-Align PointCloud to axis aligned bounding box of target PointCloud.

input PointCloud and target PointCloud should be of same size. inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud.

See also: realign, realign!

Flux3D.ReAlignTriMesh — Type

ReAlignTriMesh(target::TriMesh; inplace::Bool=true)

Re-Align TriMesh to axis aligned bounding box of mesh at index in TriMesh target.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: realign, realign!

Flux3D.RotatePointCloud — Type

RotatePointCloud(rotmat::AbstractArray{<:Number,2}; inplace::Bool=true)

Rotate PointCloud with a given rotation matrix rotmat.

rotmat must be AbstractArray{<:Number,2} of size (3,3). inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud. Given rotmat, this transform will rotate each point coordinates (ie. x,y,z) in PointCloud.

See also: rotate, rotate!

Flux3D.RotateTriMesh — Type

RotateTriMesh(rotmat::AbstractArray{<:Number,2}; inplace::Bool=true)

Rotate vertices of TriMesh with a given rotation matrix rotmat.

rotmat must be AbstractArray{<:Number,2} of size (3,3). inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh. Given rotmat, this transform will rotate each vertices coordinates (ie. x,y,z) in TriMesh.

See also: rotate, rotate!

Flux3D.ScalePointCloud — Type

ScalePointCloud(factor::Number; inplace::Bool=true)

Scale PointCloud with a given scaling factor factor.

factor should be strictly greater than 0.0 for obvious reason. inplace is optional keyword argument, to make transformation in-place. If inplace is set to false, it will create deepcopy of PointCloud. Given factor, this transform scale each point in PointCloud, ie. point = point * factor

See also: scale, scale!

Flux3D.ScaleTriMesh — Type

ScaleTriMesh(factor::Number; inplace::Bool=true)

Scale TriMesh with a given scaling factor factor.

factor should be strictly greater than 0.0 for obvious reason. inplace is optional keyword argument, to make transformation in-place. If inplace is set to false, it will create deepcopy of TriMesh. Given factor, this transform scale each vertices in TriMesh, ie. point = point * factor

See also: scale, scale!

Flux3D.TranslateTriMesh — Type

TranslateTriMesh(vector::AbstractArray{<:Number}; inplace::Bool=true)

Translate TriMesh with given translation vector.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: translate, translate!

Flux3D.TriMesh — Type

TriMesh

Initialize Triangle Mesh representation.

Fields:

N - Batch size of TriMesh.

V - Maximum vertices per mesh in TriMesh.

F - Maximum faces per mesh in TriMesh.

equalised - Bool, indicates all mesh have same verts and faces size.

valid -

offset - Offset indicating number to be added to migrate to 0-indexed system.

_verts_len - Number of vertices in each mesh of TriMesh.

_verts_list - Vertices in list format.

_verts_packed - Vertices in packed format.

_verts_padded - Vertices in padded format.

_faces_len - Number of faces in each mesh of TriMesh.

_faces_list - Vertices in list format.

_faces_packed - Vertices in packed format.

_faces_padded - Vertices in padded format.

_edges_packed - Edges in packed format (according to packed vertices).

_faces_to_edges_packed - Faces formed by edges in packed format (according to packed edges).

_laplacian_packed - Laplacian sparce matrix in packed format.

_edges_to_key - Dict mapping edges tuple to unique key.

Available Contructor:

TriMesh(verts_list, faces_list; offset::Number = -1)

TriMesh(m::Vector{<:GeometryBasics.Mesh})

TriMesh(m::GeometryBasics.Mesh)

TriMesh(m::TriMesh)

Flux3D.GBMesh — Method

GBMesh(m::TriMesh; index::Int = 1)
GBMesh(verts::AbstractArray{T,2}, faces::AbstractArray{R,2}) where {T,R}

Initialize GeometryBasics.Mesh from triangle mesh in TriMesh m at index.

See also: gbmeshes

Flux3D.compute_faces_areas_list — Method

compute_faces_areas_list(m::TriMesh)

Computes area of faces of TriMesh m in list format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: compute_faces_areas_padded, compute_faces_areas_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_faces_areas_list(m)`

Flux3D.compute_faces_areas_packed — Method

compute_faces_areas_packed(m::TriMesh)

Computes area of faces of TriMesh m in packed format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: compute_faces_areas_padded, compute_faces_areas_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_faces_areas_packed(m)`

Flux3D.compute_faces_areas_padded — Method

compute_faces_areas_padded(m::TriMesh)

Computes area of faces of TriMesh m in padded format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: compute_faces_areas_packed, compute_faces_areas_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_faces_areas_padded(m)`

Flux3D.compute_faces_normals_list — Method

compute_faces_normals_list(m::TriMesh)

Computes Unit normal of faces of TriMesh m in list format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: compute_faces_normals_padded, compute_faces_normals_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_faces_normals_list(m)`

Flux3D.compute_faces_normals_packed — Method

compute_faces_normals_packed(m::TriMesh)

Computes Unit normal of faces of TriMesh m in packed format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: compute_faces_normals_padded, compute_faces_normals_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_faces_normals_packed(m)`

Flux3D.compute_faces_normals_padded — Method

compute_faces_normals_padded(m::TriMesh)

Computes Unit normal of faces of TriMesh m in padded format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: compute_faces_normals_packed, compute_faces_normals_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_faces_normals_padded(m)`

Flux3D.compute_verts_normals_list — Method

compute_verts_normals_list(m::TriMesh)

Computes Unit normal of vertices of TriMesh m in list format.

refresh is an optional keyword argument to recompute this format for TriMesh m. Normal vec of each vertex is weighted sum of normals of adjacent faces, weighted by corresponding faces areas and then normalize to unit vector.

See also: compute_verts_normals_padded, compute_verts_normals_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_verts_normals_list(m)`

Flux3D.compute_verts_normals_packed — Method

compute_verts_normals_packed(m::TriMesh)

Computes Unit normal of vertices of TriMesh m in packed format.

refresh is an optional keyword argument to recompute this format for TriMesh m. Normal vec of each vertex is weighted sum of normals of adjacent faces, weighted by corresponding faces areas and then normalize to unit vector.

See also: compute_verts_normals_padded, compute_verts_normals_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_verts_normals_packed(m)`

Flux3D.compute_verts_normals_padded — Method

compute_verts_normals_padded(m::TriMesh)

Computes Unit normal of vertices of TriMesh m in padded format.

refresh is an optional keyword argument to recompute this format for TriMesh m. Normal vec of each vertex is weighted sum of normals of adjacent faces, weighted by corresponding faces areas and then normalize to unit vector.

See also: compute_verts_normals_packed, compute_verts_normals_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> compute_verts_normals_padded(m)`

Flux3D.gbmeshes — Method

gbmeshes(m::TriMesh)

Initialize list of GeometryBasics.Mesh from TriMesh m

See also: gbmeshes

Flux3D.get_edges_packed — Method

get_edges_packed(m::TriMesh; refresh::Bool = false)

Returns edges of TriMesh m in packed format. Edges are according to the indices of corresponding vertices in get_verts_packed(m)

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_verts_packed, get_faces_to_edges_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_edges_packed(m)`

Flux3D.get_edges_to_key — Method

get_edges_to_key(m::TriMesh; refresh::Bool = false)

Returns dict mapping edges (tuple) of TriMesh m to unique key. Edges are according to the indices of corresponding vertices inget_verts_packed(m)

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_verts_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_edges_to_key(m)`

Flux3D.get_faces_list — Method

get_faces_list(m::TriMesh; refresh::Bool = false)

Returns faces of TriMesh m in list format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_faces_padded, get_faces_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_faces_list(m)`

Flux3D.get_faces_packed — Method

get_faces_packed(m::TriMesh; refresh::Bool = false)

Returns faces of TriMesh m in packed format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_faces_padded, get_faces_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_faces_packed(m)`

Flux3D.get_faces_padded — Method

get_faces_padded(m::TriMesh; refresh::Bool = false)

Returns faces of TriMesh m in padded format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_faces_padded, get_faces_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_faces_padded(m)`

Flux3D.get_faces_to_edges_packed — Method

get_faces_to_edges_packed(m::TriMesh; refresh::Bool = false)

Returns faces of TriMesh m in form of edges. Each edge corresponds to the indices of get_edges_packed(m).

refresh is an optional keyword argument to recompute this format for TriMesh m. Assuming face f consists of (v1,v2,v3) vertices, and e1 = {v2,v3}, e2 = {v3,v1}, e3 = {v1,v2}, so face f in form of edges would be (e1,e2,e3).

See also: get_edges_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_faces_to_edges_packed(m)`

Flux3D.get_laplacian_packed — Method

get_laplacian_packed(m::TriMesh; refresh::Bool = false)
get_laplacian_sparse(m::TriMesh; refresh::Bool = false)

Returns Laplacian sparce matrix of TriMesh m.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: laplacian_loss

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_laplacian_packed(m)`

Flux3D.get_verts_list — Method

get_verts_list(m::TriMesh; refresh::Bool = false)

Returns vertices of TriMesh m in list format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_verts_padded, get_verts_packed

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_verts_list(m)`

Flux3D.get_verts_packed — Method

get_verts_packed(m::TriMesh; refresh::Bool = false)

Returns vertices of TriMesh m in packed format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_verts_padded, get_verts_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_verts_packed(m)`

Flux3D.get_verts_padded — Method

get_verts_padded(m::TriMesh; refresh::Bool = false)

Returns vertices of TriMesh m in padded format.

refresh is an optional keyword argument to recompute this format for TriMesh m.

See also: get_verts_packed, get_verts_list

Examples:

julia julia> m = load_trimesh("teapot.obj") julia> get_verts_padded(m)`

Flux3D.load_trimesh — Method

load_trimesh(fn::String)
load_trimesh(fns::Vector{String})

Load TriMesh from file(s). It will load TriMesh with multiple meshes, if list of files fns is given.

Supported formats are obj, stl, ply, off and 2DM.

Flux3D.npoints — Method

npoints(p::PointCloud)

Returns the size of PointCloud.

Flux3D.sample_points — Method

sample_points(m::TriMesh, num_samples::Int=5000; returns_normals::Bool=false, eps::Number = 1e-6)

Uniformly samples num_samples points from the surface of TriMesh m.

returns_normals is optional keyword argument, to returns normals from respective faces of samples. eps is optional keyword argument for a small number to prevent division by zero for small surface areas.

Examples:

julia> m = loadtrimesh("teapot.obj") julia> points = samplepoints(m, 5000) julia> points, normals = samplepoints(m, 5000; returnsnormals=true)

Flux3D.save_trimesh — Function

save_trimesh(fn::String, mesh::TriMesh, index::Int = 1)
save_trimesh(fn::String, mesh::GeometryBasics.Mesh)

Save mesh in given fn. index is an optional argument specifing the index of mesh, incase of multiple meshes in TriMesh mesh.

Supported formats are obj, stl, ply, off and 2DM.

Flux3D._load_meta — Method

_load_meta(m::GeometryBasics.Mesh)

Returns vertices and faces in Array format.

Flux3D._packed_to_padded — Method

packed (sum(Mi),3) padded (N, max(Mi), 3)

Flux3D.normalize! — Method

normalize!(pcloud::PointCloud)

Normalize the PointCloud pcloud with mean centered at origin and unit standard deviation and overwrite the pcloud with normalized PointCloud.

See also: normalize

Examples:

julia> p = PointCloud(rand(1024,3))
julia> normalize!(p)

Flux3D.normalize! — Method

normalize!(m::TriMesh)

Normalize each mesh in TriMesh m with mean centered at origin and unit standard deviation and overwrite the m with normalized TriMesh.

See also: normalize

Examples:

julia> m = load_trimesh("teapot.obj")
julia> normalize!(m)

Flux3D.normalize — Method

normalize(pcloud::PointCloud)

Normalize the PointCloud pcloud with mean centered at origin and unit standard deviation

See also: normalize!

Examples:

julia> p = PointCloud(rand(1024,3))
julia> p = normalize(p)

Flux3D.normalize — Method

normalize(m::TriMesh)

Normalize each mesh in TriMesh m with mean centered at origin and unit standard deviation

See also: normalize!

Examples:

julia> m = load_trimesh("teapot.obj")
julia> m = normalize(m)

Flux3D.offset — Method

offset(m::TriMesh, offset_verts_packed::AbstractArray{<:Number,2})

Add offset to the vertices of the TriMesh m by offset vertices offset_verts_packed.

See also: offset!

Examples:

```julia julia> m = loadtrimesh("teapot.obj") julia> offsetverts = ones(getvertspacked(m)) julia> m = offset(m, offset_verts)

Flux3D.realign! — Method

realign!(src::PointCloud, tgt::PointCloud)
realign!(src::PointCloud, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})

Re-Align the PointCloud src with the axis aligned bounding box of PointCloud tgt and overwrite pcloud with re-aligned PointCloud.

PointCloud src and tgt should be of same dimension.

See also: realign

Examples:

julia> src = PointCloud(rand(1024,3))
julia> tgt = PointCloud(rand(1024,3))
julia> realign!(src, tgt)

Flux3D.realign! — Method

realign!(src::TriMesh, tgt::TriMesh)
realign!(src::TriMesh, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})

Re-Align the TriMesh src with the axis aligned bounding box of mesh at index in TriMesh tgt and overwrite src with re-aligned TriMesh.

See also: realign

Examples:

julia> src = load_trimesh("teapot.obj")
julia> tgt = scale(src, 2.0)
julia> realign!(src, tgt)

Flux3D.realign — Method

realign(src::PointCloud, tgt::PointCloud)
realign(src::PointCloud, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})

Re-Align the PointCloud src with the axis aligned bounding box of PointCloud tgt.

PointCloud src and tgt should be of same dimension.

See also: realign!

Examples:

julia> src = PointCloud(rand(1024,3))
julia> tgt = PointCloud(rand(1024,3))
julia> src = realign!(src, tgt)

Flux3D.realign — Method

realign(src::TriMesh, tgt::TriMesh, index::Integer=1)
realign(src::TriMesh, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})

Re-Align the TriMesh src with the axis aligned bounding box of mesh at index in TriMesh tgt.

See also: realign

Examples:

julia> src = load_trimesh("teapot.obj")
julia> tgt = scale(src, 2.0)
julia> src = realign(src, tgt)

Flux3D.rotate! — Method

rotate!(pcloud::PointCloud, rotmat::AbstractArray{Number,2})

Rotate the PointCloud pcloud by rotation matrix rotmat and overwrite pcloud with rotated PointCloud.

Rotation matrix rotmat should be of size (3,3)

See also: rotate

Examples:

julia> p = PointCloud(rand(1024,3))
julia> rotmat = rand(3,3)
julia> rotate!(p, rotmat)

Flux3D.rotate! — Method

rotate!(m::TriMesh, rotmat::AbstractArray{<:Number,2})

Rotate the TriMesh m by rotation matrix rotmat and overwrite m with rotated TriMesh.

Rotation matrix rotmat should be of size (3,3)

See also: rotate

Examples:

julia> m = load_trimesh("teapot.obj")
julia> rotmat = rand(3,3)
julia> rotate!(m, rotmat)

Flux3D.rotate — Method

rotate(pcloud::PointCloud, rotmat::Array{Number,2})

Rotate the PointCloud pcloud by rotation matrix rotmat.

Rotation matrix rotmat should be of size (3,3)

See also: rotate!

Examples:

julia> p = PointCloud(rand(1024,3))
julia> rotmat = rand(3,3)
julia> p = rotate(p, rotmat)

Flux3D.rotate — Method

rotate(m::TriMesh, rotmat::AbstractArray{<:Number,2})

Rotate the TriMesh m by rotation matrix rotmat.

Rotation matrix rotmat should be of size (3,3)

See also: rotate!

Examples:

julia> m = load_trimesh("teapot.obj")
julia> rotmat = rand(3,3)
julia> m = rotate(m, rotmat)

Flux3D.scale! — Method

scale!(pcloud::PointCloud, factor::Number)

Scale the PointCloud pcloud by scaling factor factor and overwrite pcloud with scaled PointCloud.

Scaling factor factor should be strictly greater than 0.0.

See also: scale

Examples:

julia> p = PointCloud(rand(1024,3))
julia> scale!(p, 1.0)

Flux3D.scale! — Method

scale!(m::TriMesh, factor::Number)
scale!(m::TriMesh, factor::AbstractArray{<:Number})

Scale the TriMesh m by scaling factor factor and overwrite m with scaled TriMesh. If factor is array of size (3, ), then TriMesh will be scale by scaling factor of respective dimension.

Scaling factor factor (each element in case of array) should be strictly greater than 0.0.

See also: scale

Examples:

julia> m = load_trimesh("teapot.obj")
julia> scale!(m, 1.0)
julia> scale!(m, [1.0, 1.0, 1.0])

Flux3D.scale — Method

scale(pcloud::PointCloud, factor::Number)

Scale the PointCloud pcloud by scaling factor factor.

Scaling factor factor should be strictly greater than 0.0.

See also: scale!

Examples:

julia> p = PointCloud(rand(1024,3))
julia> p = scale(1.0)

Flux3D.scale — Method

scale(m::TriMesh, factor::Number)
scale(m::TriMesh, factor::AbstractArray{<:Number})

Scale the TriMesh m by scaling factor factor. If factor is array of size (3, ), then TriMesh will be scaleby scaling factor of respective dimension.

Scaling factor factor (each element in case of array) should be strictly greater than 0.0.

See also: scale!

Examples:

```julia julia> m = load_trimesh("teapot.obj") julia> m = scale(m, 1.0) julia> m = scale!(m, [1.0, 1.0, 1.0])

Flux3D.translate! — Method

translate!(m::TriMesh, vector::Number)
translate!(m::TriMesh, vector::AbstractArray{<:Number})

Translate the TriMesh m by translating vector vector and overwrite m with translated TriMesh. If vector is a number, then TriMesh will be translated by same number in all dimension.

See also: translate

Examples:

```julia julia> m = load_trimesh("teapot.obj") julia> translate!(m, 0.0) julia> translate!(m, [0.0, 0.0, 0.0])

Flux3D.translate — Method

translate(m::TriMesh, vector::Number)
translate(m::TriMesh, vector::AbstractArray{<:Number})

Translate the TriMesh m by translating vector vector. If vector is a number, then TriMesh will be translated by same number in all dimension.

See also: translate!

Examples:

```julia julia> m = load_trimesh("teapot.obj") julia> m = translate(m, 0.0) julia> m = translate(m, [0.0, 0.0, 0.0])