Inter-Conversion between different 3D Structure

Converting a 3D Structure to any other Structure can be achieve using the respective 3D structure constructor. We can also use transforms for inter-conversion between different structure.

Flux3D.PointCloud — Type

PointCloud(v::VoxelGrid, npoints::Int = 1000; thresh::Number = 0.5f0, algo = :MarchingCubes)

Initialize PointCloud from the VoxelGrid Structures having npoints.

thresh is the threshold from which to make binary voxels, and algo is the mode to be used to convert binary voxels. Available algo are [:Exact, :MarchingCubes, :MarchingTetrahedra, :NaiveSurfaceNets].

See also: VoxelGridToPointCloud

Flux3D.PointCloud — Type

PointCloud(m::TriMesh, npoints::Int = 1000)

Initialize PointCloud, having npoints, from the TriMesh structure.

See also: TriMeshToPointCloud

Flux3D.TriMesh — Type

TriMesh(p::PointCloud, res::Int = 32; algo = :MarchingCubes)

Initialize TriMesh from PointCloud structures having specified resolution.

algo is the mode to be used to convert binary voxels. Available algo are [:Exact, :MarchingCubes, :MarchingTetrahedra, :NaiveSurfaceNets].

See also: PointCloudToTriMesh

Flux3D.TriMesh — Method

TriMesh(v::VoxelGrid; thresh::Number = 0.5f0, algo = :MarchingCubes)

Initialize TriMesh from the VoxelGrid structures.

threshold is the threshold from which to make binary voxels, and algo is the mode to be used to convert binary voxels. Available algo are [:Exact, :MarchingCubes, :MarchingTetrahedra, :NaiveSurfaceNets].

See also: VoxelGridToTriMesh

Flux3D.VoxelGrid — Type

VoxelGrid(p::PointCloud, res::Int = 32)

Initialize VoxelGrid, having specified resolution, from PointCloud Structure.

See also: PointCloudToVoxelGrid

Flux3D.VoxelGrid — Type

VoxelGrid(m::TriMesh, res::Int = 32)

Initialize VoxelGrid, having specified resolution, from TriMesh Structure.

See also: TriMeshToVoxelGrid