Transforms

Usage

Transforms can be applied to corresponding 3D structure. Chain can be used to compose multiple functions/transforms. Same as Flux.Chain.

Chain also supports indexing and slicing. See Example below.

Example

julia> t = Chain(ScalePointCloud(2.0), NormalizePointCloud())
Chain(ScalePointCloud(factor=2.0; inplace=true), NormalizePointCloud(;inplace=true))

julia> p = PointCloud(rand(1024,3))
PointCloud{Float32} Structure:
    Batch size: 1
    Points: 3
    Normals 0
    Storage type: Array{Float32,3}

julia> t(p) == t[2](t[1](p))
true

julia> m = load_trimesh("teapot.obj")
TriMesh{Float32, UInt32, Array} Structure:
    Batch size: 1
    Max verts: 1202
    Max faces: 2256
    offset: -1
    Storage type: Array

julia> t = NormalizeTriMesh()
NormalizeTriMesh(;inplace=true)

julia> t(m)
TriMesh{Float32, UInt32, Array} Structure:
    Batch size: 1
    Max verts: 1202
    Max faces: 2256
    offset: -1
    Storage type: Array

List of all available Transforms

Flux3D.Dataset.CustomDataset

Flux3D.Dataset.ModelNet10.ModelNet10PCloud

Flux3D.Dataset.ModelNet40.ModelNet40PCloud

Flux3D.NormalizePointCloud

Flux3D.NormalizeTriMesh

Flux3D.OffsetTriMesh

Flux3D.PointCloud

Flux3D.ReAlignPointCloud

Flux3D.ReAlignTriMesh

Flux3D.RotatePointCloud

Flux3D.RotateTriMesh

Flux3D.ScalePointCloud

Flux3D.ScaleTriMesh

Flux3D.TranslateTriMesh

Flux3D.TriMesh

Flux3D.Dataset.ModelNet10.dataset

Flux3D.Dataset.ModelNet40.dataset

Flux3D.GBMesh

Flux3D.compute_faces_areas_list

Flux3D.compute_faces_areas_packed

Flux3D.compute_faces_areas_padded

Flux3D.compute_faces_normals_list

Flux3D.compute_faces_normals_packed

Flux3D.compute_faces_normals_padded

Flux3D.compute_verts_normals_list

Flux3D.compute_verts_normals_packed

Flux3D.compute_verts_normals_padded

Flux3D.gbmeshes

Flux3D.get_edges_packed

Flux3D.get_edges_to_key

Flux3D.get_faces_list

Flux3D.get_faces_packed

Flux3D.get_faces_padded

Flux3D.get_faces_to_edges_packed

Flux3D.get_laplacian_packed

Flux3D.get_laplacian_sparse

Flux3D.get_verts_list

Flux3D.get_verts_packed

Flux3D.get_verts_padded

Flux3D.load_trimesh

Flux3D.normalize

Flux3D.normalize

Flux3D.normalize!

Flux3D.normalize!

Flux3D.npoints

Flux3D.offset

Flux3D.realign

Flux3D.realign

Flux3D.realign!

Flux3D.realign!

Flux3D.rotate

Flux3D.rotate

Flux3D.rotate!

Flux3D.rotate!

Flux3D.sample_points

Flux3D.sample_points

Flux3D.save_trimesh

Flux3D.scale

Flux3D.scale

Flux3D.scale!

Flux3D.scale!

Flux3D.translate

Flux3D.translate!

Flux3D.visualize

TriMesh Transforms

Flux3D.NormalizeTriMesh — Type

NormalizeTriMesh(; inplace::Bool=true)

Normalize TriMesh with mean at origin and unit standard deviation.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: normalize, normalize!

Flux3D.ScaleTriMesh — Type

ScaleTriMesh(factor::Number; inplace::Bool=true)

Scale TriMesh with a given scaling factor factor.

factor should be strictly greater than 0.0 for obvious reason. inplace is optional keyword argument, to make transformation in-place. If inplace is set to false, it will create deepcopy of TriMesh. Given factor, this transform scale each vertices in TriMesh, ie. point = point * factor

See also: scale, scale!

Flux3D.RotateTriMesh — Type

RotateTriMesh(rotmat::AbstractArray{<:Number,2}; inplace::Bool=true)

Rotate vertices of TriMesh with a given rotation matrix rotmat.

rotmat must be AbstractArray{<:Number,2} of size (3,3). inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh. Given rotmat, this transform will rotate each vertices coordinates (ie. x,y,z) in TriMesh.

See also: rotate, rotate!

Flux3D.ReAlignTriMesh — Type

ReAlignTriMesh(target::TriMesh; inplace::Bool=true)

Re-Align TriMesh to axis aligned bounding box of mesh at index in TriMesh target.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: realign, realign!

Flux3D.TranslateTriMesh — Type

TranslateTriMesh(vector::AbstractArray{<:Number}; inplace::Bool=true)

Translate TriMesh with given translation vector.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: translate, translate!

Flux3D.OffsetTriMesh — Type

OffsetTriMesh(offset_verts::AbstractArray{<:Number,2}; inplace::Bool=true)

Add offset to the TriMesh by given offset vertices offset_verts

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of TriMesh.

See also: offset, offset!

PointCloud Transforms

Flux3D.NormalizePointCloud — Type

NormalizePointCloud(; inplace::Bool=true)

Normalize PointCloud with mean at origin and unit standard deviation.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud.

See also: normalize, normalize!

Flux3D.ScalePointCloud — Type

ScalePointCloud(factor::Number; inplace::Bool=true)

Scale PointCloud with a given scaling factor factor.

factor should be strictly greater than 0.0 for obvious reason. inplace is optional keyword argument, to make transformation in-place. If inplace is set to false, it will create deepcopy of PointCloud. Given factor, this transform scale each point in PointCloud, ie. point = point * factor

See also: scale, scale!

Flux3D.RotatePointCloud — Type

RotatePointCloud(rotmat::AbstractArray{<:Number,2}; inplace::Bool=true)

Rotate PointCloud with a given rotation matrix rotmat.

rotmat must be AbstractArray{<:Number,2} of size (3,3). inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud. Given rotmat, this transform will rotate each point coordinates (ie. x,y,z) in PointCloud.

See also: rotate, rotate!

Flux3D.ReAlignPointCloud — Type

ReAlignPointCloud(target::PointCloud; inplace::Bool=true)
ReAlignPointCloud(target::AbstractArray{<:Number, 2}; inplace::Bool=true)

Re-Align PointCloud to axis aligned bounding box of target PointCloud.

input PointCloud and target PointCloud should be of same size. inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud.

See also: realign, realign!