ModelNet

ModelNet10 Dataset

Flux3D.Dataset.ModelNet10.dataset — Method

dataset(;mode=:pointcloud, kwargs...)

Returns ModelNet10 dataset.

Supported mode are {:pointcloud}.

Optional Arguments:

For mode=:pointcloud

root::String=default_root - Root directory of dataset

train::Bool=true - Specifies the trainset

npoints::Int=1024 - Number of points and normals in each PointCloud.

transform=nothing - Transform to be applied to data point.

sampling=nothing - 'to be implement'

Examples:

julia> dset = ModelNet10.dataset(;mode=:pointcloud, train=false)
julia> typeof(dset[1].data) == PointCloud

Flux3D.Dataset.ModelNet10.ModelNet10PCloud — Type

ModelNet10PCloud

PointCloud version of ModelNet10 dataset.

Fields:

root::String - Root directory of dataset

path::String - Directory of dataset

train::Bool - Specifies the trainset

length::Int - Length of dataset.

datapaths::Array - Array containing the shape and path for each datapoint.

npoints::Int - Number of points and normals in each PointCloud.

transform::Union{Flux3D.AbstractTransform, Nothing} - - Transform to be applied to data point.

sampling::Nothing - 'to be implement'

classes_to_idx::Dict{String, UInt8} - Dict mapping from shape name to class_idx

classes::Array{String,1} - Array of shape names.

ModelNet40 Dataset

Flux3D.Dataset.ModelNet40.dataset — Method

dataset(;mode=:pointcloud, kwargs...)

Returns ModelNet40 dataset.

Supported mode are {:pointcloud}.

Optional Arguments:

For mode=:pointcloud

root::String=default_root - Root directory of dataset

train::Bool=true - Specifies the trainset

npoints::Int=1024 - Number of points and normals in each PointCloud.

transform=nothing - Transform to be applied to data point.

sampling=nothing - 'to be implement'

Examples:

julia> dset = ModelNet40.dataset(;mode=:pointcloud, train=false)
julia> typeof(dset[1].data) == PointCloud

Flux3D.Dataset.ModelNet40.ModelNet40PCloud — Type

ModelNet40PCloud

PointCloud version of ModelNet40 dataset.

Fields:

root::String - Root directory of dataset

path::String - Directory of dataset

train::Bool - Specifies the trainset

length::Int - Length of dataset.

datapaths::Array - Array containing the shape and path for each datapoint.

npoints::Int - Number of points and normals in each PointCloud.

transform::Union{Flux3D.AbstractTransform, Nothing} - - Transform to be applied to data point.

sampling::Nothing - 'to be implement'

classes_to_idx::Dict{String, UInt8} - Dict mapping from shape name to class_idx

classes::Array{String,1} - Array of shape names.