PointCloud Structure

PointCloud Constructor

Flux3D.PointCloud — Type

PointCloud

Initialize PointCloud representation.

points should be Array of size (D, N, B) where N is the number of points, D is dimensionality of each points (i.e. D=2 or D=3) and B is the batch size of PointCloud. normals is optional field, if given should be Array of size (D, N, B) where N and B should match with the N and B of points and D=2 or D=3 (i.e. normals for 2D and 3D PointCloud respectively).

Fields:

points - Points that makes up whole PointCloud.

normals - Normals of each points in PointCloud.

Available Contructor:

PointCloud(points, normals=nothing)

PointCloud(;points, normals=nothing)

PointCloud(pcloud::PointCloud)

Flux3D.npoints — Method

npoints(p::PointCloud)

Returns the size of PointCloud.