Flux3D

Flux3D.Compose

Flux3D.NormalizePointCloud

Flux3D.PointCloud

Flux3D.ReAlignPointCloud

Flux3D.RotatePointCloud

Flux3D.ScalePointCloud

Flux3D.normalize

Flux3D.normalize!

Flux3D.npoints

Flux3D.realign

Flux3D.realign!

Flux3D.rotate

Flux3D.rotate!

Flux3D.scale

Flux3D.scale!

Flux3D.visualize

Flux3D.Compose — Type

Compose(transforms...)

Composes multiple transforms/functions sequentially. Compose also support indexing and slicing.

Examples:

julia> t = Compose(x->2*x, x->x/2)
julia> t(2) == 2

julia> t = Compose(ScalePointCloud(2.0), NormalizePointCloud())
julia> p = PointCloud(rand(1024,3))
julia> t(p) == t[2](t[1](p))

Flux3D.NormalizePointCloud — Type

NormalizePointCloud(; inplace::Bool=true)

Normalize PointCloud with mean at origin and unit standard deviation.

inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud.

See also: normalize, normalize!

Flux3D.PointCloud — Type

PointCloud

Initialize PointCloud representation.

points should be Array of size (N,D) where N is the number of points and D is dimensionality of each points (i.e. D=2 or D=3). normals is optional field, if given should be Array of size (N,D) where N should match with the N of points and D=2 or D=3 (i.e. normals for 2D and 3D PointCloud respectively).

Fields:

points - Points that makes up whole PointCloud.

normals - Normals of each points in PointCloud.

Available Contructor:

PointCloud(points::Array{T,2}, normals::Union(Array{R,2}, nothing)=nothing) where {T<:Number,R<:Number}

PointCloud(;points, normals=nothing)

PointCloud(pcloud::PointCloud)

Flux3D.ReAlignPointCloud — Type

ReAlignPointCloud(target::PointCloud; inplace::Bool=true)
ReAlignPointCloud(target::AbstractArray{<:Number, 2}; inplace::Bool=true)

Re-Align PointCloud to axis aligned bounding box of target PointCloud.

input PointCloud and target PointCloud should be of same size. inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud.

See also: realign, realign!

Flux3D.RotatePointCloud — Type

RotatePointCloud(rotmat::AbstractArray{<:Number,2}; inplace::Bool=true)

Rotate PointCloud with a given rotation matrix rotmat.

rotmat must be AbstractArray{<:Number,2} of size (3,3). inplace is optional keyword argument, to make transformation in-place If inplace is set to false, it will create deepcopy of PointCloud. Given rotmat, this transform will rotate each point coordinates (ie. x,y,z) in PointCloud.

See also: rotate, rotate!

Flux3D.ScalePointCloud — Type

ScalePointCloud(factor::Number; inplace::Bool=true)

Scale PointCloud with a given scaling factor factor.

factor should be strictly greater than 0.0 for obvious reason. inplace is optional keyword argument, to make transformation in-place. If inplace is set to false, it will create deepcopy of PointCloud. Given factor, this transform scale each point in PointCloud, ie. point = point * factor

See also: scale, scale!

Flux3D.npoints — Method

npoints(p::PointCloud)

Returns the size of PointCloud.

Flux3D.visualize — Method

visualize(pcloud::PointCloud; kwargs...)

Visualize PointCloud pcloud.

Dimension of points in PointCloud pcloud must be 3.

Optional Arguments:

color (Symbol) - Color of the marker, default :blue

markersize (Number) - Size of the marker, default 0.02*npoints(pcloud)/1024

Flux3D.normalize! — Method

normalize!(pcloud::PointCloud)

Normalize the PointCloud pcloud with mean centered at origin and unit standard deviation and overwrite the pcloud with normalized PointCloud.

See also: normalize

Examples:

julia> p = PointCloud(rand(1024,3))
julia> normalize!(p)

Flux3D.normalize — Method

normalize(pcloud::PointCloud)

Normalize the PointCloud pcloud with mean centered at origin and unit standard deviation

See also: normalize!

Examples:

julia> p = PointCloud(rand(1024,3))
julia> p = normalize(p)

Flux3D.realign! — Method

realign!(src::PointCloud, tgt::PointCloud)
realign!(src::PointCloud, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})

Re-Align the PointCloud src with the axis aligned bounding box of PointCloud tgt and overwrite pcloud with re-aligned PointCloud.

PointCloud src and tgt should be of same dimension.

See also: realign

Examples:

julia> src = PointCloud(rand(1024,3))
julia> tgt = PointCloud(rand(1024,3))
julia> realign!(src, tgt)

Flux3D.realign — Method

realign(src::PointCloud, tgt::PointCloud)
realign(src::PointCloud, tgt_min::AbstractArray{<:Number,2}, tgt_max::AbstractArray{<:Number,2})

Re-Align the PointCloud src with the axis aligned bounding box of PointCloud tgt.

PointCloud src and tgt should be of same dimension.

See also: realign!

Examples:

julia> src = PointCloud(rand(1024,3))
julia> tgt = PointCloud(rand(1024,3))
julia> src = realign!(src, tgt)

Flux3D.rotate! — Method

rotate!(pcloud::PointCloud, rotmat::AbstractArray{Number,2})

Rotate the PointCloud pcloud by rotation matrix rotmat and overwrite pcloud with rotated PointCloud.

Rotation matrix rotmat should be of size (3,3)

See also: rotate

Examples:

julia> p = PointCloud(rand(1024,3))
julia> rotmat = rand(3,3)
julia> rotate!(p, rotmat)

Flux3D.rotate — Method

rotate(pcloud::PointCloud, rotmat::Array{Number,2})

Rotate the PointCloud pcloud by rotation matrix rotmat.

Rotation matrix rotmat should be of size (3,3)

See also: rotate!

Examples:

julia> p = PointCloud(rand(1024,3))
julia> rotmat = rand(3,3)
julia> p = rotate(p, rotmat)

Flux3D.scale! — Method

scale!(pcloud::PointCloud, factor::Number)

Scale the PointCloud pcloud by scaling factor factor and overwrite pcloud with scaled PointCloud.

Scaling factor factor should be strictly greater than 0.0.

See also: scale

Examples:

julia> p = PointCloud(rand(1024,3))
julia> scale!(p, 1.0)

Flux3D.scale — Method

scale(pcloud::PointCloud, factor::Number)

Scale the PointCloud pcloud by scaling factor factor.

Scaling factor factor should be strictly greater than 0.0.

See also: scale!

Examples:

julia> p = PointCloud(rand(1024,3))
julia> p = scale(1.0)